# docs/swagger.yaml
schemes:
  - http
swagger: '2.0'
info:
  description: server API for Question Board Application.
  title: Question Board
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
host: localhost:8080
basePath: /api
paths:
  /login:
    post:
      description: Login and get Authorization Token
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Authorization
      summary: Login
      parameters:
        - description: send request email, password
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/request.LoginRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.UserDetails'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
...
  /admin/major:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of Major, for Administrator only.
      tags:
        - Major
      summary: Get All Major
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/models.Major'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
    post:
      security:
        - ApiKey: []
      description: Route Path for Insert New Major, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Major
      summary: Create New Major
      parameters:
        - description: send request major code and major name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.Major'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/major/{id}/delete:
    delete:
      security:
        - ApiKey: []
      description: Route Path for Delete Major, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Major
      summary: Delete Major
      parameters:
        - type: integer
          description: major id
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/major/{id}/update:
    put:
      security:
        - ApiKey: []
      description: Route Path for Update Major, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Major
      summary: Update Major
      parameters:
        - type: integer
          description: major id
          name: id
          in: path
          required: true
        - description: send request major code and major name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.Major'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
  /refresh_token:
    post:
      description: Route Path for Get New Access Token
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Authorization
      summary: Refresh Token
      parameters:
        - description: send request access_token, refresh_token
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.Token'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/models.Token'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
  /student/evaluate:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of Evaluation Quest with Subject ID and Class.
      tags:
        - Evaluate
      summary: Get Evaluate Quest
      parameters:
        - description: send user data
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/request.User'
        - type: integer
          description: subject id
          name: subject_id
          in: query
          required: true
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    $ref: '#/definitions/response.Quest'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
    post:
      security:
        - ApiKey: []
      description: Route Path for Answer Evaluation Quest.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Evaluate
      summary: Create New Quest
      parameters:
        - type: string
          description: class
          name: class
          in: query
        - type: integer
          description: teacher id
          name: teacher_id
          in: query
        - type: integer
          description: subject id
          name: subject_id
          in: query
        - description: send quest data
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/request.Answer'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
  /teacher/evaluate:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get Quest Response By Quest ID.
      tags:
        - Evaluate
      summary: Quest Response By Quest ID
      parameters:
        - type: integer
          description: subject id
          name: subject_id
          in: query
          required: true
        - type: integer
          description: class
          name: class
          in: query
          required: true
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    $ref: '#/definitions/response.QuestResponses'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
definitions:
  models.AnswerOption:
    type: object
    properties:
      id:
        type: integer
      question_id:
        type: integer
      string_answer:
        type: string
  models.EvaluateTeacher:
    type: object
    properties:
      class:
        type: string
      id:
        type: integer
      subject:
        $ref: '#/definitions/models.Subject'
      subject_id:
        type: integer
      teacher_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
  models.Level:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  models.Major:
    type: object
    properties:
      code:
        type: string
        example: INF
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Informatics
  models.Question:
    type: object
    properties:
      answer_option:
        type: array
        items:
          $ref: '#/definitions/models.AnswerOption'
      id:
        type: integer
      question:
        type: string
      questionnaire_id:
        type: integer
      user_response:
        type: array
        items:
          $ref: '#/definitions/models.UserAnswer'
      with_option:
        type: integer
  models.Questionnaire:
    type: object
    properties:
      completor:
        type: array
        items:
          $ref: '#/definitions/models.User'
      createdAt:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: integer
      questions:
        type: array
        items:
          $ref: '#/definitions/models.Question'
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
  models.Subject:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      major_id:
        type: integer
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/models.User'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/models.TeacherSubject'
  models.Tag:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Year
      value:
        type: string
        example: '2019'
  models.TeacherSubject:
    type: object
    properties:
      class:
        type: string
      subject:
        $ref: '#/definitions/models.Subject'
      subject_id:
        type: integer
      teacher:
        $ref: '#/definitions/models.User'
      teacher_id:
        type: integer
  models.Token:
    type: object
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
  models.User:
    type: object
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      level_id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      major_id:
        type: integer
      name:
        type: string
      password:
        type: string
      status:
        type: integer
      subject:
        type: array
        items:
          $ref: '#/definitions/models.Subject'
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      teacher_subject:
        type: array
        items:
          $ref: '#/definitions/models.TeacherSubject'
      updatedAt:
        type: string
  models.UserAnswer:
    type: object
    properties:
      answer:
        type: string
      evaluateTeacher:
        $ref: '#/definitions/models.EvaluateTeacher'
      evaluate_teacher_id:
        type: integer
      id:
        type: integer
      question_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
  request.Answer:
    type: object
    properties:
      answers:
        type: array
        items:
          $ref: '#/definitions/request.UserAnswer'
      quest:
        $ref: '#/definitions/models.Questionnaire'
      user:
        $ref: '#/definitions/models.User'
  request.LoginRequest:
    type: object
    properties:
      email:
        type: string
        example: admin@web.io
      password:
        type: string
        example: admin
...
  request.User:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      name:
        type: string
      status:
        type: integer
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
  request.UserAnswer:
    type: object
    properties:
      answer:
        type: string
      id:
        type: integer
      question_id:
        type: integer
...
  response.Error:
    type: object
    properties:
      error: {}
      message:
        type: string
  response.MessageData:
    type: object
    properties:
      data: {}
      message:
        type: string
  response.MessageOnly:
    type: object
    properties:
      message:
        type: string
  response.Quest:
    type: object
    properties:
      description:
        type: string
      id:
        type: integer
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      title:
        type: string
  response.QuestResponses:
    type: object
    properties:
      id:
        type: integer
      questions:
        type: array
        items:
          $ref: '#/definitions/response.Respondent'
      title:
        type: string
  response.Respondent:
    type: object
    properties:
      id:
        type: integer
      number_of_response:
        type: integer
      question:
        type: string
      user_response:
        type: array
        items:
          $ref: '#/definitions/response.UserAnswer'
  response.Student:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
  response.Subject:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/response.Student'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/response.TeacherClass'
  response.SubjectTeacher:
    type: object
    properties:
      class:
        type: string
      subject:
        $ref: '#/definitions/response.SubjectWithoutTeacher'
...
  response.SubjectWithoutTeacher:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
  response.Teacher:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  response.TeacherClass:
    type: object
    properties:
      class:
        type: string
      teacher:
        $ref: '#/definitions/response.Teacher'
  response.User:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      name:
        type: string
      status:
        type: integer
  response.UserAnswer:
    type: object
    properties:
      answer:
        type: string
      id:
        type: integer
      question_id:
        type: integer
      user:
        $ref: '#/definitions/response.User'
  response.UserDetails:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      status:
        type: integer
      subject:
        type: array
        items:
          $ref: '#/definitions/response.Subject'
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      teacher_subject:
        type: array
        items:
          $ref: '#/definitions/response.SubjectTeacher'
securityDefinitions:
  ApiKey:
    type: apiKey
    name: Authorization
    in: header